requirements.txt

fastapi[standard]==0.119.1
requests==2.32.5
cachetools==6.2.1
python-dotenv==1.2.1